refactor: split GameLogic.ts
