+++ server/src/models/index.ts
@@ -1,6 +1,8 @@
 import { Player } from "./Player";
 import { Room } from "./Room";
 import { RoomManager } from "./RoomManager";
-import { Timer, Game, TransitionCondition, PhasesTransitionsConditions, PhaseRouter } from "./GameLogic";
+import { TransitionCondition, PhasesTransitionsConditions, PhaseRouter } from "./PhaseRouter";
+import { Timer } from "./Timer";
+import {Game} from "./Game"
 
 export { Player, Room, RoomManager, Timer, Game, TransitionCondition, PhasesTransitionsConditions, PhaseRouter };
